Step 3: Controls
Now it is time to build our first little UI by replacing the “Hello
World” text in the HTML body by the SAPUI5 control sap.m.Text. In the
beginning, we will use the JavaScript control interface to set up the
UI, the control instance is then placed into the HTML body.