Step 7: JSON Model
Now that we have set up the view and controller, it’s about time to
think about the "M" in MVC. We will add an input field to our app, bind
its value to the model, and bind the same value to the description of
the input field. The description will be directly updated as the user
types.