Step 8: Translatable Texts
In this step we move the texts of our UI to a separate resource file.
This way, they are all in a central place and can be easily translated
into other languages. This process of internationalization – in short
i18n – is achieved in SAPUI5 by using a special resource model and the
standard data binding syntax, but without preceding /.