Step 9: Component Configuration
After we have introduced all three parts of the Model-View-Controller
(MVC) concept, we now come to another important structural aspect of
SAPUI5. In this step, we will encapsulate all UI assets in a component
that is independent from our index.html file. Components are independent
and reusable parts used in SAPUI5 applications. Whenever we access
resources, we will now do this relatively to the component (instead of
relatively to the index.html). This architectural change allows our app
to be used in more flexible environments than our static index.html
page, such as in a surrounding container like the SAP Fiori launchpad.