Step 10: Descriptor for Applications
All application-specific configuration settings will now further be put
in a separate descriptor file called manifest.json. This clearly
separates the application coding from the configuration settings and
makes our app even more flexible. For example, all SAP Fiori
applications are realized as components and come with a descriptor file
in order to be hosted in the SAP Fiori launchpad.

The SAP Fiori launchpad acts as an application container and
instantiates the app without having a local HTML file for the bootstrap.
Instead, the descriptor file will be parsed and the component is loaded
into the current HTML page. This allows several apps to be displayed in
the same context. Each app can define local settings, such as language
properties, supported devices, and more. And we can also use the
descriptor file to load additional resources and instantiate models like
our i18n resource bundle.