Step 13: Margins and Paddings
Our app content is still glued to the corners of the letterbox. To
fine-tune our layout, we can add margins and paddings to the controls
that we added in the previous step. But instead of manually adding CSS
to the controls, we will use the standard classes provided by SAPUI5.
These classes take care of consistent sizing steps, left-to-right
support, and responsiveness.