Step 19: Reuse Dialogs
In step 16 we created a dialog as fragment, because we wanted it to be
reusable across views or across our whole app. But we placed the logic
for retrieving the dialog instance and for opening and closing it
respectively in the controller of the HelloPanel view. Sticking to this
approach would require copying and pasting the code to the controller of
each view that needs our dialog. This would of course cause an undesired
code redundancy we definitely want to avoid. In this step, we will
implement the solution to this problem: We now expand our reuse concept
and invoke the dialog at component level.