Step 21: Data Types
The list of invoices is already looking nice, but what is an invoice
without a price assigned? Typically prices are stored in a technical
format and with a '.' delimiter in the data model. For example, our
invoice for pineapples has the calculated price 87.2 without a currency.
We are going to use the SAPUI5 data types to format the price properly,
with a locale-dependent decimal separator and two digits after the
separator.