Step 26: Remote OData Service
So far we have only worked with local JSON data, but now we will access
a real OData service. Instead of implementing an own OData service we
will simply use the publicly available Northwind OData service to
visualize remote data. You will be surprised how little needs to be
changed in order to make this work!