Step 28: Unit Test with QUnit
Now that we have a test folder in the app, we can start to increase our
test coverage. Actually, every feature that we added to the app so far,
would require a separate test case. We have totally neglected this so
far, so let’s add a simple unit test for our custom formatter function
from Step 23. We will test if the long text for our status is correct by
comparing it with the texts from our resource bundle.