Step 29: Integration Test with OPA
If we want to test interaction patterns or more visual features of our
app, we can also write an integration test. We haven’t thought about
testing our interaction with the app yet, so in this step we will check
if the dialog actually opens when we click the “Say Hello with Dialog”
button. We can easily do this with OPA5, a feature of SAPUI5 that is
easy to set up and is based on JavaScript and QUnit. Using integration
and unit tests and running them consistently in a continuous integration
(CI) environment, we can make sure that we don’t accidentally break our
app or introduce logical errors in existing code.