Step 30: Debugging Tools
Even though we have added a basic test coverage in the previous steps,
it seems like we accidentally broke our app, because it does not display
prices to our invoices anymore. We need to debug the issue and fix it
before someone finds out. Luckily, SAPUI5 provides a couple of debugging
tools that we can use within the app to check the application logic and
the developer tools of modern browsers are also quite good. We will now
check for the root cause.