Step 31: Routing and Navigation
So far, we have put all app content on one single page. As we add more
and more features, we want to split the content and put it on separate
pages. In this step, we will use the SAPUI5 navigation features to load
and show a separate detail page that we can later use to display details
for an invoice. In the previous steps, we defined the page directly in
the app view so that it is displayed when the app is loaded. We will now
use the SAPUI5 router class to load the pages and update the URL for us
automatically. We specify a routing configuration for our app and create
a separate view for each page of the app, then we connect the views by
triggering navigation events.