Step 32: Routing with Parameters
We can now navigate between the overview and the detail page, but the
actual item that we selected in the overview is not displayed on the
detail page yet. A typical use case for our app is to show additional
information for the selected item on the detail page. To make this work,
we have to pass over the information which item has been selected to the
detail page and show the details for the item there.