Step 34: Custom Controls
In this step, we are going to extend the functionality of SAPUI5 with a
custom control. We want to rate the product shown on the detail page, so
we create a composition of multiple standard controls using the SAPUI5
extension mechanism and add some glue code to make them work nicely
together. This way, we can reuse the control across the app and keep all
related functionality in one module.