Step 35: Responsiveness
In this step, we improve the responsiveness of our app. SAPUI5
applications can be run on phone, tablet, and desktop devices and we can
configure the application to make best use of the screen estate for each
scenario. Fortunately, SAPUI5 controls like the sap.m.Table already
deliver a lot of features that we can use.