Step 37: Content Density
In the last step of our Walkthrough tutorial, we adjust the content
density based on the user’s device. SAPUI5 contains different content
densities allowing you to display larger controls for touch-enabled
devices and a smaller, more compact design for devices that are operated
by mouse. In our app, we will detect the device and adjust the density
accordingly.